fix(APPC-4170): remove ask notification permission from home
