+++ app/src/main/java/com/asfoundation/wallet/wallet/home/HomeFragment.kt
@@ -1,13 +1,11 @@
 package com.asfoundation.wallet.wallet.home
 
-import android.os.Build
 import android.os.Bundle
 import android.preference.PreferenceManager
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -97,9 +95,6 @@ class HomeFragment : BasePageViewFragment(), SingleStateFragment<HomeState, Home
   private val navBarViewModel: NavBarViewModel by activityViewModels()
   private val hasGetSomeValidBalanceResult = mutableStateOf(false)
 
-  private val pushNotificationPermissionLauncher =
-    registerForActivityResult(ActivityResultContracts.RequestPermission()) {}
-
   override fun onCreateView(
     inflater: LayoutInflater,
     container: ViewGroup?,
@@ -111,7 +106,6 @@ class HomeFragment : BasePageViewFragment(), SingleStateFragment<HomeState, Home
   override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
     super.onViewCreated(view, savedInstanceState)
     viewModel.collectStateAndEvents(lifecycle, viewLifecycleOwner.lifecycleScope)
-    askNotificationsPermission()
   }
 
   override fun onResume() {
@@ -359,15 +353,6 @@ class HomeFragment : BasePageViewFragment(), SingleStateFragment<HomeState, Home
 
       is HomeSideEffect.NavigateToBalanceDetails ->
         navigator.navigateToBalanceBottomSheet(sideEffect.balance)
-    }
-  }
-
-  private fun askNotificationsPermission() {
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-      // Android OS manages when to ask for permission. After android 11, the default behavior is
-      // to only prompt for permission if needed, and if the user didn't deny it twice before. If
-      // the user just dismissed it without denying, then it'll prompt again next time.
-      pushNotificationPermissionLauncher.launch(android.Manifest.permission.POST_NOTIFICATIONS)
     }
   }
